// src/hrd/src/app/model/profile.model.ts
export class ProfileModel {
   id: string;
   nik: string;
   nama: string;
   jenkel: string;
   tempatlahir: string;
   tgllahir: string;
   agama: string;
   status: string;
   alamat: string;
   foto: string;
   jabatan: string;
   statuskaryawan: string;
   tglmasuk: string;
   masakerja: string;
   pendidikan: string;
   email: string;
   notelp: string;

}